requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.0
pydantic==2.9.2
pandas==2.2.3
polars==1.3.0
yfinance==0.2.40
requests==2.32.3
redis==5.0.7
openpyxl==3.1.5
XlsxWriter==3.2.0
python-dotenv==1.0.1
orjson==3.10.7

